moved getting the current date code into getCurrentDate Function, added/modified comments, changed variable name for current_date
