+++ models/routes/routes.js
@@ -2,11 +2,20 @@
 
 const { localStorage } = require('node-localstorage');
 const router = require('express').Router();
-//import todo model
-const todoItemsModel = require('../mongodb/todoTaks');
+const todoItemsModel = require('../mongodb/todoTask');
 const bodyParser = require('body-parser');
 
-//create first route --add Todo Item to database
+function getCurrentDate() {
+  let ts = Date.now();
+  let date_ob = new Date(ts);
+  let date = date_ob.getDate();
+  let month = date_ob.getMonth() + 1;
+  let year = date_ob.getFullYear();
+  let currentDate = month + '/' + date + '/' + year;
+  return currentDate;
+}
+
+//Add Todo Item to the database
 router.post('/todos', async (req, res) => {
   try {
     const newItem = new todoItemsModel({
@@ -14,8 +23,6 @@ router.post('/todos', async (req, res) => {
       date: req.body.date,
       email: req.body.email,
     });
-
-    //save this item in database
     const saveItem = await newItem.save();
     res.status(200).json(saveItem);
   } catch (err) {
@@ -23,52 +30,36 @@ router.post('/todos', async (req, res) => {
   }
 });
 
-//create second route -- get data from database
+//Get data from the database
 router.get('/todos', async (req, res) => {
   try {
-    let ts = Date.now();
-
-    let date_ob = new Date(ts);
-    let date = date_ob.getDate();
-    let month = date_ob.getMonth() + 1;
-    let year = date_ob.getFullYear();
-    let current_date = month + '/' + date + '/' + year;
+    let currentDate = getCurrentDate();
     console.log('EMAIL' + req.query.email);
-    //console.log("EMAIL"+localStorage.getItem('email'))
     const allTodoItems = await todoItemsModel.find({
-      $and: [{ email: req.query.email, date: current_date }],
+      $and: [{ email: req.query.email, date: currentDate }],
     });
 
     console.log(allTodoItems);
-    // prints date & time in YYYY-MM-DD format
-    //const allTodoItems =  todoItemsModel.find().populate({match: { date: "2022-11-21"}}).exec();
     res.status(200).json(allTodoItems);
   } catch (err) {
     res.json(err);
   }
 });
+
+//Get data for tomorrow
 router.get('/todostomorrow', async (req, res) => {
   try {
-    let ts = Date.now();
-
-    let date_ob = new Date(ts);
-    let date = date_ob.getDate();
-    let month = date_ob.getMonth() + 1;
-    let year = date_ob.getFullYear();
-    let current_date = month + '/' + date + '/' + year;
+    let currentDate = getCurrentDate();
     const allTodoItems = await todoItemsModel
-      .find({ email: req.query.email, date: { $ne: current_date } })
+      .find({ email: req.query.email, date: { $ne: currentDate } })
       .sort({ date: 1 });
-
-    // prints date & time in YYYY-MM-DD format
-    //const allTodoItems =  todoItemsModel.find().populate({match: { date: "2022-11-21"}}).exec();
     res.status(200).json(allTodoItems);
   } catch (err) {
     res.json(err);
   }
 });
 
-//update item
+//Find an item by its ID and update it
 router.put('/todos', async (req, res) => {
   if (req.body.item != null) {
     todoItemsModel.findByIdAndUpdate(
@@ -86,6 +77,8 @@ router.put('/todos', async (req, res) => {
     );
   }
 });
+
+//Mark a todo item as complete
 router.get('/todo/complete/:id', async (req, res) => {
   const todo = await todoItemsModel.findById(req.params.id);
   if (todo != null) {
@@ -95,10 +88,9 @@ router.get('/todo/complete/:id', async (req, res) => {
   }
 });
 
-//Delete item from database
+//Find the item by its id and delete it from the database
 router.delete('/todos/:id', async (req, res) => {
   try {
-    //find the item by its id and delete it
     console.log('IN DELETE ID IS ' + req.params.id);
     const deleteItem = await todoItemsModel.findByIdAndDelete(req.params.id);
     res.status(200).json('Item Deleted');
@@ -108,5 +100,4 @@ router.delete('/todos/:id', async (req, res) => {
   }
 });
 
-//export router
 module.exports = router;
